+++ src/features/outfits/outfit-feed/outfit-post/outfit-post-slider/index.ts
@@ -1,0 +1,3 @@
+import OutfitPostSlider from './outfit-post-slider';
+
+export default OutfitPostSlider;
